requirement.txt
openai
langchain
streamlit
huggingface_hub
python-dotenv
streamlit